+++ internal/db/querier.go
@@ -10,8 +10,8 @@ import (
 
 type Querier interface {
 	CreateLink(ctx context.Context, arg CreateLinkParams) (Link, error)
-	DeleteLink(ctx context.Context, id int32) (int64, error)
-	GetLinkById(ctx context.Context, id int32) (Link, error)
+	DeleteLink(ctx context.Context, id int64) (int64, error)
+	GetLinkById(ctx context.Context, id int64) (Link, error)
 	GetLinkByshortName(ctx context.Context, shortName string) (Link, error)
 	GetLinks(ctx context.Context) ([]Link, error)
 	UpdateLink(ctx context.Context, arg UpdateLinkParams) (Link, error)
